Defined files for style only and turned them into tsx
