+++ src/components/Particle/Particle.tsx
@@ -18,11 +18,11 @@ const Particle = () => {
               enable: true,
               mode: "repulse",
             },
-            onresize: {
+/*            onresize: {
               enable: true,
               density_auto: true,
               density_area: 300 // nb_particles = particles.nb * (canvas width *  canvas height / 1000) / density_area
-            },
+            }, */
             resize: true,
           },
           modes: {
