+++ src/components/TypeEffect/styles.tsx
@@ -1,0 +1,13 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles((theme) => ({
+    text: {
+        bottom: "2.5rem",
+        position: "relative",
+        fontSize: "1.5rem",
+        color:"white",
+        fontFamily: 'monospace'
+    },
+}));
+
+export default useStyles;
