+++ tsconfig.json
@@ -22,7 +22,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx"
-  ],
+, "src/components/Articles/styles.tsx", "src/pages/aboutMe.js", "src/pages/contact.js"  ],
   "exclude": [
     "node_modules"
   ]
